Port BlockEvent, PlayerInteractEvent  to neoforge1.21.6

+++ neoforge-compat/src/main/java/xiao/battleroyale/compat/neoforge/event/NeoFarmlandTrampleEvent.java
@@ -56,7 +56,7 @@ public class NeoFarmlandTrampleEvent extends NeoEvent implements IFarmlandTrampl
 
     @Override
     public float getFallDistance() {
-        return farmlandTrampleEvent.getFallDistance();
+        return (float) farmlandTrampleEvent.getFallDistance(); // ?为什么改成double 还有高手?!
     }
 
     @Override
